Better header rendering for a white bg
